On 64-bit systems, void* is 8 bytes long. So cast as long.
git-archimport-id: sbromley@pangeosubsea.com--pangeo-2007a/tclargv--sam-dev--1.0--patch-2

+++ src/tclArgv.c
@@ -163,7 +163,7 @@ gotMatch:
 		infoPtr = matchPtr;
 		switch (infoPtr->type) {
 			case TCL_ARGV_CONSTANT:
-				*((int *) infoPtr->dst) = (int) infoPtr->src;
+				*((int *) infoPtr->dst) = (int) ((long)infoPtr->src);
 				break;
 			case TCL_ARGV_INT:
 				if (objc == 0) {
